Move cholesky api to tensor.py. test=release/2.0
