+++ python/paddle/tensor/linalg.py
@@ -25,7 +25,7 @@ __all__ = [
     'dist',
     't',
     'cross',
-    'cholesky',
+    #  'cholesky',
     #  'tensordot',
     'bmm'
 ]
@@ -600,59 +600,6 @@ def cross(input, other, dim=None):
                 'Y': other},
         outputs={'Out': out},
         attrs=attrs)
-    return out
-
-
-def cholesky(x, upper=False):
-    """
-    Computes the Cholesky decomposition of one symmetric positive-definite
-    matrix or batches of symmetric positive-definite matrice. 
-    
-    If `upper` is `True`, the decomposition has the form :math:`A = U^{T}U` ,
-    and the returned matrix :math:`U` is upper-triangular. Otherwise, the
-    decomposition has the form  :math:`A = LL^{T}` , and the returned matrix
-    :math:`L` is lower-triangular.
-
-    Args:
-        x (Variable): The input tensor. Its shape should be `[*, M, M]`,
-            where * is zero or more batch dimensions, and matrices on the
-            inner-most 2 dimensions all should be symmetric positive-definite.
-            Its data type should be float32 or float64.
-        upper (bool): The flag indicating whether to return upper or lower
-            triangular matrices. Default: False.
-
-    Returns:
-        Variable: A Tensor with same shape and data type as `x`. It represents \
-            triangular matrices generated by Cholesky decomposition.
-        
-    Examples:
-        .. code-block:: python
-
-            import paddle
-            import paddle.fluid as fluid
-            import numpy as np
-
-            with fluid.dygraph.guard():
-                a = np.random.rand(3, 3)
-                a_t = np.transpose(a, [1, 0])
-                x = np.matmul(a, a_t) + 1e-03
-                x = fluid.dygraph.to_variable(x)
-                out = paddle.cholesky(x, upper=False)
-                print(out.numpy())
-                # [[1.190523   0.         0.        ]
-                #  [0.9906703  0.27676893 0.        ]
-                #  [1.25450498 0.05600871 0.06400121]]
-
-    """
-    check_variable_and_dtype(x, 'dtype', ['float32', 'float64'], 'cholesky')
-    check_type(upper, 'upper', bool, 'cholesky')
-    helper = LayerHelper('cholesky', **locals())
-    out = helper.create_variable_for_type_inference(dtype=x.dtype)
-    helper.append_op(
-        type='cholesky',
-        inputs={'X': [x]},
-        outputs={'Out': out},
-        attrs={'upper': upper})
     return out
